+++ CHANGELOG.md
@@ -15,9 +15,11 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Gracefully shutdown all concurrent tasks by forwarding the SIGTERM signal ([#788]).
 - Bump testing-tools to `0.3.0-stackable0.0.0-dev` ([#780]).
 
 [#780]: https://github.com/stackabletech/druid-operator/pull/780
+[#788]: https://github.com/stackabletech/druid-operator/pull/788
 
 ## [25.11.0] - 2025-11-07